proyecto/redes/requirements.txt
backports.csv==1.0.7
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1
django-crum==0.7.3
django-import-export==1.2.0
django-jet==1.0.8
et-xmlfile==1.0.1
jdcal==1.4
mpmath==1.1.0
numpy==1.16.1
odfpy==1.4.0
openpyxl==2.6.1
#pkg-resources==0.0.0
pytz==2018.9
PyYAML==5.1
scipy==1.2.1
sympy==1.3
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
